refacator: move `Slug` from `domain_entities` to `value_objects` module
