+++ src/domain/domain_entities/article.rs
@@ -2,9 +2,9 @@ use chrono::NaiveDateTime as DateTime;
 use uuid::Uuid;
 
 use super::article_tag::ArticleTag;
-use super::slug::Slug;
 use super::user::User;
 use crate::domain::value_objects::changeset::{BlankChangeSet, ChangeSet, Changes};
+use crate::domain::value_objects::slug::Slug;
 use crate::libs::time::TimeHelper;
 use crate::util::{verify_role_has_permission, RolePermissions};
