+++ src/domain/services/journalism/comments/comment_on_article_service.rs
@@ -62,7 +62,7 @@ mod test {
     use super::*;
     use crate::domain::domain_entities::article::Article;
     use crate::domain::domain_entities::article_tag::DraftArticleTag;
-    use crate::domain::domain_entities::slug::Slug;
+    use crate::domain::value_objects::slug::Slug;
     use crate::domain::repositories::article_tag_repository::ArticleTagRepositoryTrait;
     use crate::libs::time::TimeHelper;
     use crate::tests::relationship_managers::comment_article::CommentArticleRelationInMemoryManager;
