+++ src/domain/value_objects/mod.rs
@@ -1,2 +1,3 @@
 pub mod changeset;
 pub mod count_summary;
+pub mod slug;
