+++ src/domain/value_objects/slug.rs
